requirements.txt
alembic==1.8.0
black==22.6.0
click==8.1.3
Flask==2.1.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.1
MarkupSafe==2.1.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
psycopg2-binary==2.9.3
python-dotenv==0.20.0
SQLAlchemy==1.4.39
tomli==2.0.1
typing_extensions==4.3.0
Werkzeug==2.1.2
zipp==3.8.0
